Move api models to api folder
