+++ src/models/api/WorkerPoolRequest.ts
