+++ src/models/index.ts
@@ -11,6 +11,10 @@ export * from './api/NotificationResponse'
 export * from './api/NotificationUpdateRequest'
 export * from './api/SchemaResponse'
 export * from './api/StateUpdateRequest'
+export * from './api/WorkerFlowRunResponse'
+export * from './api/WorkerFlowRunsRequest'
+export * from './api/WorkerPoolRequest'
+export * from './api/WorkerPoolResponse'
 export * from './api/WorkQueueHealthPolicyResponse'
 export * from './api/WorkQueueStatusResponse'
 export * from './ApiRouteParams'
@@ -81,15 +85,11 @@ export * from './TaskRunResponse'
 export * from './UiFlowRunHistory'
 export * from './UiFlowRunHistoryResponse'
 export * from './WorkerFlowRun'
-export * from './WorkerFlowRunResponse'
 export * from './WorkerFlowRuns'
-export * from './WorkerFlowRunsRequest'
 export * from './WorkerPool'
 export * from './WorkerPoolCreate'
 export * from './WorkerPoolEdit'
 export * from './WorkerPoolFilter'
-export * from './WorkerPoolRequest'
-export * from './WorkerPoolResponse'
 export * from './WorkQueue'
 export * from './WorkQueueCreate'
 export * from './WorkQueueEdit'
